Desafios clases 17 y 18

+++ CursoCoder/Assets/Script/Player/PlayerData.cs
@@ -9,8 +9,8 @@ public class PlayerData : ScriptableObject
     [SerializeField] private float speed;
     public float Speed{get{ return speed; }}
 
-    [SerializeField] private float hp;
-    public float HP { get{ return hp; }  set { hp = value; } }
+    [SerializeField] private int hp;
+    public int HP { get{ return hp; }  set { hp = value; } }
 
     [SerializeField] private int score = 0;
     public int Score { get{ return score; } set{ score = value; } }
